# openapi/valk-dashboard.yaml
openapi: 3.1.0
info:
  title: VALK Dashboard API
  version: 2.0.0
  description: Contracts consumed by the Next.js BFF. Legacy API-key authentication remains supported beside short-lived dashboard bearer tokens.
servers:
  - url: /api
security:
  - dashboardBearer: []
  - legacyApiKey: []
paths:
...
  /summary/monthly-performance/assessment:
    post:
      operationId: createMonthlyAssessment
      x-capability: assessment:run
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                {
                  schema:
                    {
                      type: object,
                      required: [monthly_data],
                      properties:
                        {
                          monthly_data:
                            {
                              type: array,
                              items:
                                { type: object, additionalProperties: true },
                            },
                        },
                    },
                },
            },
        }
      responses:
        {
          "200":
            {
              description: Assessment result,
              content:
                {
                  application/json:
                    { schema: { type: object, additionalProperties: true } },
                },
            },
          default: { $ref: "#/components/responses/ErrorResponse" },
        }
  /summary/discord/report:
    post:
      operationId: sendDiscordReport
      x-capability: reports:send
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                {
                  schema:
                    {
                      type: object,
                      required: [page],
                      properties:
                        {
                          page: { type: string },
                          mode: { type: string, enum: [full, top5] },
                        },
                    },
                },
            },
        }
      responses:
        {
          "200":
            {
              description: Report status,
              content:
                {
                  application/json:
                    { schema: { type: object, additionalProperties: true } },
                },
            },
          default: { $ref: "#/components/responses/ErrorResponse" },
        }
...
  /colonisation/contributions:
    get:
      operationId: getColonisationContributions
      responses:
        {
          "200": { $ref: "#/components/responses/DataResponse" },
          default: { $ref: "#/components/responses/ErrorResponse" },
        }
  /colonisation/constructions:
    get:
      operationId: getColonisationConstructions
      responses:
        {
          "200": { $ref: "#/components/responses/DataResponse" },
          default: { $ref: "#/components/responses/ErrorResponse" },
        }
...
  /account/discord-webhook:
    get:
      operationId: getPersonalDiscordWebhookStatus
      security: [{ dashboardBearer: [] }]
      x-capability: rules:write
      responses:
        "200": { description: Masked webhook configuration status }
        default: { $ref: "#/components/responses/ErrorResponse" }
    put:
      operationId: setPersonalDiscordWebhook
      security: [{ dashboardBearer: [] }]
      x-capability: rules:write
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [webhook_url]
              properties:
                { webhook_url: { type: string, format: uri, maxLength: 2048 } }
      responses:
        "200": { description: Webhook encrypted and saved }
        default: { $ref: "#/components/responses/ErrorResponse" }
    delete:
      operationId: deletePersonalDiscordWebhook
      security: [{ dashboardBearer: [] }]
      x-capability: rules:write
      responses:
        "200": { description: Webhook removed }
        default: { $ref: "#/components/responses/ErrorResponse" }
...
  /table/{table}:
    get:
      operationId: getTableRows
      x-capability: admin:read
      parameters:
        - { name: table, in: path, required: true, schema: { type: string } }
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Sort"
        - $ref: "#/components/parameters/Direction"
        - {
            name: filters,
            in: query,
            schema: { type: string },
            description: JSON-encoded structured filter expression,
          }
      responses:
        {
          "200": { $ref: "#/components/responses/DataResponse" },
          default: { $ref: "#/components/responses/ErrorResponse" },
        }
...
components:
  securitySchemes:
    dashboardBearer: { type: http, scheme: bearer, bearerFormat: JWT }
    legacyApiKey: { type: apiKey, in: header, name: apikey }
  parameters:
    Page:
      {
        name: page,
        in: query,
        schema: { type: integer, minimum: 1, default: 1 },
      }
    PageSize:
      {
        name: page_size,
        in: query,
        schema: { type: integer, minimum: 1, maximum: 250, default: 25 },
      }
    Sort: { name: sort, in: query, schema: { type: string } }
    Direction:
      {
        name: direction,
        in: query,
        schema: { type: string, enum: [asc, desc], default: asc },
      }
  schemas:
...
    Pagination:
      {
        type: object,
        required: [page, page_size, total],
        properties:
          {
            page: { type: integer },
            page_size: { type: integer },
            total: { type: integer },
          },
      }
    DataEnvelope:
      {
        type: object,
        required: [data, generated_at],
        properties:
          {
            data:
              {
                type: array,
                items: { type: object, additionalProperties: true },
              },
            generated_at: { type: string, format: date-time },
            pagination: { $ref: "#/components/schemas/Pagination" },
          },
      }
    ErrorEnvelope:
      {
        type: object,
        required: [error],
        properties:
          {
            error:
              {
                type: object,
                required: [code, message, correlation_id],
                properties:
                  {
                    code: { type: string },
                    message: { type: string },
                    correlation_id: { type: string },
                  },
              },
            generated_at: { type: string, format: date-time },
          },
      }
  responses:
    DataResponse:
      {
        description: Paginated data envelope,
        content:
          {
            application/json:
              { schema: { $ref: "#/components/schemas/DataEnvelope" } },
          },
      }
    ErrorResponse:
      {
        description: Uniform error envelope,
        content:
          {
            application/json:
              { schema: { $ref: "#/components/schemas/ErrorEnvelope" } },
          },
      }